fs: fuse: Allow always boost
When ht_fuse_boost is 1, if the current uid corresponds to a
foreground process, then boost is allowed.
When ht_fuse_boost is 2, if the current uid corresponds to a
system process or a foreground process, then boost is allowed.

However, the is_fg function that checks whether the process
corresponding to the current uid is foreground always checks whether
the uid is equal to -555, which is incorrect. -555 is just a dummy
value, and no uid will be -555.
The correct approach is that the userspace will get the uid of
foreground process and write it to `/proc/fg_info/fg_uids`, and the
is_fg function will receive it and make a correct judgment. Writing
to `/proc/fg_info/fg_uids` requires userspace to do.

So, fuse short circuit is not universal. Fortunately, ht_fuse_boost
is 2 by default, which means that the system process will still be
boosted normally, which will still bring some benefits.

So, let's be more aggressive. If the ht_fuse_boost value is 3, boost
is always enabled.

Also improved debug message text.