Fix snprintf buffer size in htc_recv.c
This commit fixes a bug in the htc_recv.c file where the buffer size passed to the snprintf function was incorrect. The size of the integer byteOffset was being passed instead of the size of the string byteOffsetStr. This was causing a compilation error with the -Wfortify-source flag. The buffer size for byteOffsetStr has been corrected to accommodate the hexadecimal representation of byteOffset.